+++ src/common/utils/orgUtils.ts
@@ -10,6 +10,7 @@ import * as sortArray from "sort-array";
 import { Connection, SfdxError } from "@salesforce/core";
 import { importData } from "./dataUtils";
 import { soqlQuery } from "./apiUtils";
+import { isSfdxProject } from "./projectUtils";
 
 export async function listProfiles(conn: any) {
   if (conn in [null, undefined]) {
@@ -172,7 +173,7 @@ export async function promptOrg(commandThis: any, options: any = { devHub: false
     const setDefaultUsernameCommand =
       `sfdx config:set ` +
       `${options.devHub ? "defaultdevhubusername" : "defaultusername"}=${org.username}` +
-      (!fs.existsSync(path.join(process.cwd(), "sfdx-project.json")) ? " --global" : "");
+      (!isSfdxProject() ? " --global" : "");
     await execSfdxJson(setDefaultUsernameCommand, commandThis, {
       fail: true,
       output: false,