+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
 
 - Replace use of sfpowerkit by default command `sfdx force:source:manifest:create`
 - Manage cache for listing orgs
+- Update hardis:package:version:create to allow to 
+  - install it later on an org
+  - immediately delete it
 
 ## [2.94.3] 2022-09-15